+++ genghis.rb
@@ -407,8 +407,7 @@ module Genghis
       end
 
       def collections
-        # TODO: should I be rejecting all of these `system*` collections?
-        @collections ||= @database.collections.map { |c| Collection.new(c) unless c.name.start_with? 'system' }.compact
+        @collections ||= @database.collections.map { |c| Collection.new(c) }
       end
 
       def [](coll_name)